4.1.2

+++ build/scripts/Update-Metadata.ps1
@@ -16,18 +16,25 @@ foreach ($path in $MEDIAmoviefolders) {
     foreach ($ext in $extensions) {
         $files = Get-ChildItem -LiteralPath $path -Filter $ext -Recurse
         foreach ($file in $files) {
-            $comment = (Update-Lastindex -DataSource $datasource).newcomment
             $fullname = $file.fullname
-            $oldname = $file.fullname + ".old"
-            Rename-Item $fullname $oldname -Verbose
-            ffmpeg -i $oldname -map 0:v:0? -map 0:a? -map 0:s? -metadata title="" -metadata description="" -metadata COMMENT=$comment -c copy $fullname
-            Remove-Item -LiteralPath $oldname -Force -Confirm:$false -Verbose
+            $Probe = ffprobe -loglevel 0 -print_format json -show_format $fullname
+            $convert = $Probe | ConvertFrom-Json
+            $filecomment = $convert.format.tags.comment
 
-            # If file already existed in the database and this is used due to a migration run upgrading to v4+ then update the database with the new comment.
-            $TableName = 'Movies'
-            $modified = Get-Date -Format "yyyy-MM-dd HH:mm:ss"
-            $query = "Update $TableName set comment = `"$comment`", updatedby = 'Update-Metadata', modified = `"$modified`" WHERE fullname = `"$fullname`""
-            Invoke-SqliteQuery -DataSource $DataSource -Query $query
+            if ($filecomment -like "dta-*") {
+                $fullname = $file.fullname
+                $oldname = $file.fullname + ".old"
+                $newcomment = (Update-Lastindex -DataSource $datasource).newcomment
+                Rename-Item $fullname $oldname -Verbose
+                ffmpeg -i $oldname -map 0:v:0? -map 0:a? -map 0:s? -metadata title="" -metadata description="" -metadata COMMENT=$newcomment -c copy $fullname
+                Remove-Item -LiteralPath $oldname -Force -Confirm:$false -Verbose
+
+                # If file already existed in the database and this is used due to a migration run upgrading to v4+ then update the database with the new comment.
+                $TableName = 'Movies'
+                $modified = Get-Date -Format "yyyy-MM-dd HH:mm:ss"
+                $query = "Update $TableName set comment = `"$newcomment`", updatedby = 'Update-Metadata', modified = `"$modified`" WHERE fullname = `"$fullname`""
+                Invoke-SqliteQuery -DataSource $DataSource -Query $query
+            }
         }
     }
 }
@@ -38,18 +45,25 @@ foreach ($path in $MEDIAshowfolders) {
     foreach ($ext in $extensions) {
         $files = Get-ChildItem -LiteralPath $path -Filter $ext -Recurse
         foreach ($file in $files) {
-            $comment = (Update-Lastindex -DataSource $datasource).newcomment
             $fullname = $file.fullname
-            $oldname = $file.fullname + ".old"
-            Rename-Item $fullname $oldname -Verbose
-            ffmpeg -i $oldname -map 0:v:0? -map 0:a? -map 0:s? -metadata title="" -metadata description="" -metadata COMMENT=$comment -c copy $fullname
-            Remove-Item -LiteralPath $oldname -Force -Confirm:$false -Verbose
+            $Probe = ffprobe -loglevel 0 -print_format json -show_format $fullname
+            $convert = $Probe | ConvertFrom-Json
+            $filecomment = $convert.format.tags.comment
 
-            # If file already existed in the database and this is used due to a migration run upgrading to v4+ then update the database with the new comment.
-            $TableName = 'Shows'
-            $modified = Get-Date -Format "yyyy-MM-dd HH:mm:ss"
-            $query = "Update $TableName set comment = `"$comment`", updatedby = 'Update-Metadata', modified = `"$modified`" WHERE fullname = `"$fullname`""
-            Invoke-SqliteQuery -DataSource $DataSource -Query $query
+            if ($filecomment -like "dta-*") {
+                $newcomment = (Update-Lastindex -DataSource $datasource).newcomment
+                $fullname = $file.fullname
+                $oldname = $file.fullname + ".old"
+                Rename-Item $fullname $oldname -Verbose
+                ffmpeg -i $oldname -map 0:v:0? -map 0:a? -map 0:s? -metadata title="" -metadata description="" -metadata COMMENT=$newcomment -c copy $fullname
+                Remove-Item -LiteralPath $oldname -Force -Confirm:$false -Verbose
+
+                # If file already existed in the database and this is used due to a migration run upgrading to v4+ then update the database with the new comment.
+                $TableName = 'Shows'
+                $modified = Get-Date -Format "yyyy-MM-dd HH:mm:ss"
+                $query = "Update $TableName set comment = `"$newcomment`", updatedby = 'Update-Metadata', modified = `"$modified`" WHERE fullname = `"$fullname`""
+                Invoke-SqliteQuery -DataSource $DataSource -Query $query
+            }
         }
     }
 }
